requirements.txt
dnspython == 2.5.0
